+++ lib/assorted_layout_widgets.dart
@@ -9,6 +9,7 @@ export "src/normalized_overflow_box.dart";
 export "src/pad.dart";
 export "src/row_super.dart";
 export "src/show_dialog_super.dart";
+export "src/side_by_side.dart";
 export "src/text_one_line.dart";
 export "src/text_one_line_ellipsis_with_fade.dart";
 export "src/time_builder.dart";